Split database tests by concern
